refactor(hooks): wave 3 lint hygiene for core hooks
